Absorb `AuxPieceCache` into `AuxRecordStorage`
