+++ crates/subspace-service/src/dsn.rs
@@ -1,8 +1,7 @@
 mod piece_record_store;
 
-use crate::dsn::piece_record_store::{AuxRecordStorage, SegmentIndexGetter};
+pub(crate) use crate::dsn::piece_record_store::AuxRecordStorage;
 use futures::{Stream, StreamExt};
-pub(crate) use piece_record_store::piece_cache::AuxPieceCache;
 use sc_client_api::AuxStore;
 use sc_consensus_subspace::ArchivedSegmentNotification;
 use sp_core::traits::SpawnNamed;
@@ -39,9 +38,8 @@ pub struct DsnConfig {
 
 pub(crate) async fn create_dsn_instance<Block, AS>(
     dsn_config: DsnConfig,
-    piece_cache: AuxPieceCache<AS>,
+    record_storage: AuxRecordStorage<AS>,
     piece_getter: PieceGetter,
-    segment_index_getter: SegmentIndexGetter,
 ) -> Result<
     (
         Node,
@@ -53,9 +51,6 @@ where
     Block: BlockT,
     AS: AuxStore + Sync + Send + 'static,
 {
-    // TODO: Combine `AuxPieceCache` with `AuxRecordStorage` and remove `PieceCache` abstraction
-    let record_storage = AuxRecordStorage::new(piece_cache, segment_index_getter);
-
     trace!("Subspace networking starting.");
 
     let networking_config = subspace_networking::Config {
